Version: 0.1.149 Branch: master Date: 02.07.2017 Author: AlexMerlyan
Working: RoomTypeController and RateController(roomtype service)
Changes: RoomTypeController and RateController(roomtype service)
Next:

Comment:

+++ src/main/java/projectpackage/controllers/RoomTypeController.java
@@ -7,15 +7,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import projectpackage.dto.IUDAnswer;
 import projectpackage.dto.JacksonMappingMarker;
+import projectpackage.dto.RoomTypeDTO;
 import projectpackage.model.auth.User;
 import projectpackage.model.rooms.RoomType;
-import projectpackage.dto.IUDAnswer;
-import projectpackage.service.MessageBook;
 import projectpackage.service.roomservice.RoomTypeService;
+import projectpackage.service.support.ServiceUtils;
 
-import javax.cache.annotation.CacheRemoveAll;
-import javax.cache.annotation.CacheResult;
 import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
@@ -29,12 +28,15 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 @RestController
 @RequestMapping("/roomTypes")
 public class RoomTypeController {
+
     @Autowired
     RoomTypeService roomTypeService;
 
+    @Autowired
+    ServiceUtils serviceUtils;
+
     //Get RoomType List
     @ResponseStatus(HttpStatus.OK)
-    @CacheResult(cacheName = "roomTypeList")
     @RequestMapping(method = RequestMethod.GET, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     public List<Resource<RoomType>> getRoomTypeList(){
         List<RoomType> roomTypes = roomTypeService.getAllRoomTypes();
@@ -47,17 +49,13 @@ public class RoomTypeController {
         return resources;
     }
 
-    //Get RoomType List
     @JsonView(JacksonMappingMarker.List.class)
     @ResponseStatus(HttpStatus.OK)
-    @CacheResult(cacheName = "simpleRoomTypeList")
     @RequestMapping(value = "/simpleList", method = RequestMethod.GET, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     public List<RoomType> getSimpleRoomTypeList(){
         return roomTypeService.getAllRoomTypes();
     }
 
-    //Get single RoomType by id
-    @ResponseStatus(HttpStatus.ACCEPTED)
     @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     public ResponseEntity<Resource<RoomType>> getRoomType(@PathVariable("id") Integer id, HttpServletRequest request){
         User thisUser = (User) request.getSession().getAttribute("USER");
@@ -65,9 +63,7 @@ public class RoomTypeController {
         Resource<RoomType> resource = new Resource<>(roomType);
         HttpStatus status;
         if (null != roomType){
-            if (thisUser.getRole().getRoleName().equals("ADMIN")) resource.add(linkTo(methodOn(RoomTypeController.class).deleteRoomType(roomType.getObjectId())).withRel("delete"));
-            resource.add(linkTo(methodOn(RoomTypeController.class).updateRoomType(roomType.getObjectId(), roomType)).withRel("update"));
-            status = HttpStatus.ACCEPTED;
+            status = HttpStatus.OK;
         } else {
             status = HttpStatus.BAD_REQUEST;
         }
@@ -76,43 +72,58 @@ public class RoomTypeController {
     }
 
     //Create roomType, fetch into database
-    @CacheRemoveAll(cacheName = "roomTypeList")
     @RequestMapping(method = RequestMethod.POST, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE}, consumes = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> createRoomType(@RequestBody RoomType newRoomType){
-        IUDAnswer result = roomTypeService.insertRoomType(newRoomType);
+    public ResponseEntity<IUDAnswer> createRoomType(@RequestBody RoomTypeDTO roomTypeDTO, HttpServletRequest request){
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkSessionAdminAndData(user, roomTypeDTO);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
+        }
+        RoomType roomType = new RoomType();
+        roomType.setRoomTypeTitle(roomTypeDTO.getRoomTypeTitle());
+        roomType.setContent(roomTypeDTO.getRoomTypeContent());
+        IUDAnswer result = roomTypeService.insertRoomType(roomType);
         HttpStatus status;
         if (result.isSuccessful()) {
-            status = HttpStatus.CREATED;
+            status = HttpStatus.OK;
         } else status = HttpStatus.BAD_REQUEST;
         ResponseEntity<IUDAnswer> responseEntity = new ResponseEntity<IUDAnswer>(result, status);
         return responseEntity;
     }
 
     //Update roomType method
-    @CacheRemoveAll(cacheName = "roomTypeList")
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE}, consumes = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> updateRoomType(@PathVariable("id") Integer id, @RequestBody RoomType changedRoomType){
-        if (!id.equals(changedRoomType.getObjectId())){
-            return new ResponseEntity<IUDAnswer>(new IUDAnswer(id, false, MessageBook.WRONG_UPDATE_ID), HttpStatus.NOT_ACCEPTABLE);
+    public ResponseEntity<IUDAnswer> updateRoomType(@PathVariable("id") Integer id, @RequestBody RoomTypeDTO roomTypeDTO, HttpServletRequest request){
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkSessionAdminAndData(user, roomTypeDTO);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
         }
-        IUDAnswer result = roomTypeService.updateRoomType(id, changedRoomType);
+        RoomType roomType = new RoomType();
+        roomType.setObjectId(id);
+        roomType.setContent(roomTypeDTO.getRoomTypeContent());
+        roomType.setRoomTypeTitle(roomTypeDTO.getRoomTypeTitle());
+        IUDAnswer result = roomTypeService.updateRoomType(id, roomType);
         HttpStatus status;
         if (result.isSuccessful()) {
-            status = HttpStatus.ACCEPTED;
+            status = HttpStatus.OK;
         } else status = HttpStatus.BAD_REQUEST;
-        //Creating simple ResponseEntity
         ResponseEntity<IUDAnswer> responseEntity = new ResponseEntity<IUDAnswer>(result, status);
         return responseEntity;
     }
 
     //Delete roomType method
-    @CacheRemoveAll(cacheName = "roomTypeList")
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> deleteRoomType(@PathVariable("id") Integer id){
+    public ResponseEntity<IUDAnswer> deleteRoomType(@PathVariable("id") Integer id, HttpServletRequest request){
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkDeleteForAdmin(user, id);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
+        }
         IUDAnswer result = roomTypeService.deleteRoomType(id);
         HttpStatus status;
         if (result.isSuccessful()) {
-            status = HttpStatus.ACCEPTED;
+            status = HttpStatus.OK;
         } else status = HttpStatus.NOT_FOUND;
         ResponseEntity<IUDAnswer> responseEntity = new ResponseEntity<IUDAnswer>(result, status);
         return responseEntity;
